+++ packages/air_quality/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0
+
+- Updates Kotlin plugin and AGP
+- Upgrade of compileSdkVersion
+- Upgrade of android embeddings
+
 ## 2.0.0
 
 - Migration to null safety